% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIARinterpolation.R
\name{CIARinterpolation}
\alias{CIARinterpolation}
\title{Interpolation from CIAR model}
\usage{
CIARinterpolation(
  x,
  y,
  t,
  delta = 0,
  yini = 0,
  zero.mean = TRUE,
  standardized = TRUE,
  c = 1,
  seed = 1234
)
}
\arguments{
\item{x}{An array with the parameters of the CIAR model. The elements of the array are, in order, the real (phiR) and the imaginary (phiI) part of the coefficient of CIAR model.}

\item{y}{Array with the time series observations.}

\item{t}{Array with the irregular observational times.}

\item{delta}{Array with the measurements error standard deviations.}

\item{yini}{a single value, initial value for the estimation of the missing value of the time series.}

\item{zero.mean}{logical; if TRUE, the array y has zero mean; if FALSE, y has a mean different from zero.}

\item{standardized}{logical; if TRUE, the array y is standardized; if FALSE, y contains the raw time series.}

\item{c}{Nuisance parameter corresponding to the variance of the imaginary part.}

\item{seed}{a single value, interpreted as the seed of the random process.}
}
\value{
A list with the following components:
\itemize{
\item{fitted}{ Estimation of a missing value of the CIAR process.}
\item{ll}{ Value of the negative log likelihood evaluated in the fitted missing values.}
}
}
\description{
Interpolation of missing values from models fitted by \code{\link{CIARkalman}}
}
\examples{
n=100
set.seed(6714)
st<-gentime(n)
x=CIARsample(n=n,phiR=0.9,phiI=0,st=st,c=1)
y=x$y
y1=y/sd(y)
ciar=CIARkalman(y=y1,t=st)
ciar
napos=10
y0=y1
y1[napos]=NA
xest=c(ciar$phiR,ciar$phiI)
yest=CIARinterpolation(xest,y=y1,t=st)
yest$fitted
mse=(y0[napos]-yest$fitted)^2
print(mse)
plot(st,y,type='l',xlim=c(st[napos-5],st[napos+5]))
points(st,y,pch=20)
points(st[napos],yest$fitted*sd(y),col="red",pch=20)
}
\references{
\insertRef{Elorrieta_2019}{iAR}
}
\seealso{
\code{\link{gentime}}, \code{\link{CIARsample}}, \code{\link{CIARkalman}}
}
